+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix issues when trying to munch with self hosted proxy.
 - Ravenloft subclass parsing: Rogue, Sorcerer, Warlock.
 - Dhampir 2024 parsing fixes.
+- Fix 2024 monster replacement in Native Adventure Muncher
 
 # 7.3.1